assignment3/requirements.txt
PyJWT>=2.3.0
Flask-JSON>=0.3.4
psycopg2-binary>=2.9.1
bcrypt>=3.2.0
